requirements.txt
Django
gunicorn
requests
pydub
natsort
opencv-python
psycopg2-binary
pycryptodome
pillow
djangorestframework
bson
openpyxl
mediapipe